divisi i controller per viste e api
