+++ app/Http/Controllers/WebControllers/UserController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\WebControllers;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\User;
 use Intervention\Image\Facades\Image;
@@ -12,42 +13,6 @@ use Illuminate\Support\Facades\Input;
 
 class UserController extends Controller
 {
-    ////////////////////////////////////////////////////////////////////////////////////
-    //                                    API                                         //
-    ////////////////////////////////////////////////////////////////////////////////////
-    public function index(){
-        return  User::all();
-    }
-
-    public function show ($id){
-        return User::findOrFail($id);
-    }
-
-    // lo store non lo facciamo lato api... 
-
-    public function update(Request $request, $id){
-        $user = User::findOrFail($id);
-        if ($request->hasFile('avatar')) {
-            $avatar_file = $request->file('avatar');
-            $user->updateAvatar($avatar_file);
-        }
-        if ($request->input('name')){
-            $user->name = $request->input('name');
-            $user->save();
-        }
-        return $user;
-    }
-
-    public function delete(Request $request, $id){
-        $user = User::findOrFail($id);
-        $user->delete();
-        return 204;
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////////
-    //                                    WEB                                         //
-    ////////////////////////////////////////////////////////////////////////////////////
-
     public function profile (){
         return view('profile');
     }
